move dialogue from init into own module
